MVC setup for Project; Static Menu Page
