+++ .idea/runConfigurations/Main.xml
@@ -1,0 +1,9 @@
+<component name="ProjectRunConfigurationManager">
+    <configuration default="false" name="Main" type="Application" factoryName="Application">
+        <option name="MAIN_CLASS_NAME" value="com.example.napolinights.Main" />
+        <module name="Napoli-Nights" />
+        <method v="2">
+            <option name="Make" enabled="true" />
+        </method>
+    </configuration>
+</component>
