--- src/main/java/com/example/napolinights/EmptyCartErrorController.java
@@ -1,19 +1,0 @@
-package com.example.napolinights;
-
-import javafx.fxml.FXML;
-import javafx.stage.Stage;
-import javafx.scene.control.Button;
-
-public class EmptyCartErrorController {
-
-    @FXML
-    private Button closeButton;
-
-    // Method to close the dialog
-    @FXML
-    private void handleClose() {
-        // Close the stage when the close button is pressed
-        Stage stage = (Stage) closeButton.getScene().getWindow();
-        stage.close();
-    }
-}
