--- src/main/java/com/example/napolinights/MenuController.java
@@ -1,11 +1,0 @@
-package com.example.napolinights;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Hyperlink;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.VBox;
-
-public class MenuController {
-
-
-}
