+++ src/main/java/com/example/napolinights/model/Category.java
@@ -8,6 +8,7 @@ public enum Category {
     PASTA,
     PIZZA,
     DESSERT,
+    BEVERAGES,
     NON_ALCOHOLIC_DRINKS,
     COCKTAILS
 }
