+++ src/main/java/module-info.java
@@ -10,6 +10,4 @@ module com.example.napolinights {
     opens com.example.napolinights.model to javafx.fxml;
     exports com.example.napolinights.controller;
     opens com.example.napolinights.controller to javafx.fxml;
-    exports com.example.napolinights.model.obsolete;
-    opens com.example.napolinights.model.obsolete to javafx.fxml;
 }
