--- src/main/resources/com/example/napolinights/Checkout.fxml
@@ -1,92 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.image.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
-
-<AnchorPane prefWidth="697.0" style="-fx-background-color: #141415;" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.napolinights.CheckoutController">
-    <children>
-        <!-- VBox to Center Image and Buttons -->
-        <VBox alignment="CENTER" layoutY="-86.0" prefWidth="697.0" spacing="20" AnchorPane.bottomAnchor="0" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0" AnchorPane.topAnchor="0">
-            <!-- Image Box with Background Image -->
-            <ImageView fx:id="logoImageView" fitHeight="112.0" fitWidth="200.0" preserveRatio="true" style="-fx-border-color: red; -fx-border-width: 10px; -fx-padding: 10px;">
-                <image>
-                    <Image url="@images/landing_page/napoli_nights_logo.png" />
-                </image>
-            </ImageView>
-            <HBox fx:id="imageBox" alignment="CENTER" prefHeight="297.0" prefWidth="697.0">
-            <VBox prefHeight="255.0" prefWidth="697.0" style="-fx-background-color: FFFFFF;">
-               <children>
-                  <HBox prefHeight="100.0" prefWidth="200.0" style="-fx-padding: 1em;">
-                     <children>
-                        <VBox prefHeight="76.0" prefWidth="698.0">
-                           <children>
-                              <Text strokeType="OUTSIDE" strokeWidth="0.0" style="-fx-font-weight: BOLD;" text="Cart">
-                                 <font>
-                                    <Font size="20.0" />
-                                 </font></Text>
-                              <ListView fx:id="cartListView" minHeight="-Infinity" prefHeight="88.0" prefWidth="673.0" />
-                           </children>
-                        </VBox>
-                     </children></HBox>
-                  <HBox alignment="BOTTOM_CENTER" prefHeight="149.0" prefWidth="697.0">
-                     <children>
-                        <VBox prefHeight="133.0" prefWidth="474.0">
-                           <children>
-                              <GridPane alignment="CENTER_LEFT" prefHeight="100.0" prefWidth="356.0">
-                                <columnConstraints>
-                                  <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-                                  <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-                                </columnConstraints>
-                                <rowConstraints>
-                                  <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-                                  <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-                                  <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-                                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-                                </rowConstraints>
-      
-      <!--                            CART OVERVIEW GRID-->
-                                 <children>
-                                    <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Cart items" />
-                                    <Text layoutX="10.0" layoutY="32.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Subtotal" GridPane.rowIndex="1" />
-                                    <Text layoutX="10.0" layoutY="32.0" strokeType="OUTSIDE" strokeWidth="0.0" text="GST (10%)" GridPane.rowIndex="2" />
-                                    <Text layoutX="10.0" layoutY="77.0" strokeType="OUTSIDE" strokeWidth="0.0" style="-fx-font-weight: BOLD;" text="Total" GridPane.rowIndex="3" />
-      <!--                               Quantity -->
-                                    <Text fx:id="quantityText" strokeType="OUTSIDE" strokeWidth="0.0" text=" items" GridPane.columnIndex="1" />
-      <!--                               Subtotal -->
-                                    <Text fx:id="subtotalPriceText" layoutX="188.0" layoutY="28.0" strokeType="OUTSIDE" strokeWidth="0.0" text="\$" GridPane.columnIndex="1" GridPane.rowIndex="1" />
-      <!--                              Tax -->
-                                     <Text fx:id="taxText" layoutX="188.0" layoutY="52.0" strokeType="OUTSIDE" strokeWidth="0.0" text="\$" GridPane.columnIndex="1" GridPane.rowIndex="2" />
-      <!--                              Total -->
-                                     <Text fx:id="totalPriceText" layoutX="188.0" layoutY="78.0" strokeType="OUTSIDE" strokeWidth="0.0" text="\$" GridPane.columnIndex="1" GridPane.rowIndex="3" />
-                                 </children>
-                              </GridPane>
-                              <Button fx:id="checkoutBackButton" mnemonicParsing="false" text="Back">
-                                 <VBox.margin>
-                                    <Insets left="5.0" top="15.0" />
-                                 </VBox.margin></Button>
-                           </children>
-                        </VBox>
-                        <Button fx:id="checkoutPayButton" alignment="BOTTOM_RIGHT" mnemonicParsing="false" text="Pay at counter" onMouseClicked="#handleCheckoutPayButtonClick">
-                           <HBox.margin>
-                              <Insets bottom="1.0" left="1.0" right="1.0" top="1.0" />
-                           </HBox.margin>
-                           <font>
-                              <Font size="17.0" />
-                           </font>
-                           <padding>
-                              <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
-                           </padding>
-                        </Button>
-                     </children>
-                  </HBox>
-               </children>
-            </VBox>
-
-            </HBox>
-
-        </VBox>
-    </children>
-</AnchorPane>
