--- src/main/resources/com/example/napolinights/ItemDetailDialog.fxml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.image.ImageView?>
-
-<VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" spacing="10" alignment="CENTER" fx:controller="com.example.napolinights.ItemDetailDialogController">
-    <ImageView fx:id="dialogImageView" fitWidth="300" preserveRatio="true"/>
-    <Label fx:id="dialogTitle" style="-fx-font-size: 22px;"/>
-    <Label fx:id="dialogDescription" style="-fx-font-size: 14px; -fx-text-fill: gray;"/>
-    <Label fx:id="dialogPrice" style="-fx-font-size: 18px;"/>
-
-    <!-- Quantity Controls -->
-    <HBox spacing="10" alignment="CENTER">
-        <Button text="-" onAction="#decrementQuantity"/>
-        <Label fx:id="lblQuantity" text="1"/>
-        <Button text="+" onAction="#incrementQuantity"/>
-    </HBox>
-
-    <!-- Notes TextArea -->
-    <Label text="Notes"/>
-    <TextArea fx:id="notesTextArea" prefRowCount="3"/>
-
-    <!-- Footer Buttons -->
-    <HBox spacing="20" alignment="CENTER">
-        <Button fx:id="btnCancel" text="Cancel" onAction="#handleCancel"/>
-        <Button text="Add to Cart" onAction="#handleAddToCart"/>
-    </HBox>
-</VBox>
-
