--- src/main/resources/com/example/napolinights/OrderConfirmation.fxml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.shape.Line?>
-<?import javafx.scene.shape.Rectangle?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.control.ListView?>
-<?import javafx.scene.text.Text?>
-
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/22" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.napolinights.OrderConfirmationController">
-   <children>
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="#141412" height="127.0" layoutX="-24.0" layoutY="-6.0" stroke="#ffffff00" strokeType="INSIDE" width="653.0" />
-      <ImageView fx:id="logoImageView" fitHeight="121.0" fitWidth="230.0" layoutX="237.0" layoutY="-2.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="@images/landing_page/napoli_nights_logo.png" />
-         </image>
-      </ImageView>
-      <Label fx:id="tableLabel" layoutX="300.0" layoutY="134.0" text="Table 1" />
-      <Label fx:id="OrderConfirmationLabel" layoutX="11.0" layoutY="129.0" text="Order Confirmation">
-         <font>
-            <Font size="20.0" />
-         </font>
-      </Label>
-      <VBox fx:id="OrderDetails" layoutX="13.0" layoutY="158.0" prefHeight="49.0" prefWidth="195.0">
-         <children>
-            <Label fx:id="orderNumberLabel" text="Order Number:" />
-            <Label prefHeight="17.0" prefWidth="193.0" text="Payment Method: Pay At Counter" />
-         </children>
-      </VBox>
-      <Button fx:id="LandingPageButton" layoutX="15.0" layoutY="366.0" onAction="#handleHomePage" prefHeight="0.0" prefWidth="86.0" text="Home Page" />
-
-      <!-- Food Confirmation Details Section -->
-      <VBox fx:id="FoodConfirmationDetails" alignment="CENTER" layoutX="14.0" layoutY="200.0" prefHeight="152.0" prefWidth="601.0">
-         <children>
-            <AnchorPane>
-               <children>
-                  <Rectangle arcHeight="5.0" arcWidth="5.0" fill="WHITE" height="152.0" stroke="BLACK" strokeType="INSIDE" width="602.0" />
-                  <Label layoutX="382.0" layoutY="6.0" text="Price" />
-                  <Label layoutX="561.0" layoutY="6.0" text="Total" />
-                  <Label layoutX="14.0" layoutY="6.0" text="Product" />
-                  <Label layoutX="466.0" layoutY="6.0" text="Quantity" />
-                  <Line endX="602.0" endY="26.5" startY="26.5" />
-                  <Line endX="588.0" layoutX="14.0" layoutY="125.0" startX="-14.0" />
-                  <Label layoutX="443.0" layoutY="130.0" prefHeight="17.0" prefWidth="107.0" text="Total Price (incl.):" />
-
-                  <!-- ListView for confirmed cart details (From Maria's Code) -->
-                  <ListView fx:id="confirmedCartListView" layoutY="27.0" prefHeight="99.0" prefWidth="602.0" />
-
-                  <!-- Text element for total price (From Maria's Code) -->
-                  <Text fx:id="totalPriceText" layoutX="538.0" layoutY="143.0" strokeType="OUTSIDE" strokeWidth="0.0" text="\$" wrappingWidth="47.0" />
-               </children>
-            </AnchorPane>
-         </children>
-      </VBox>
-   </children>
-</AnchorPane>
