--- src/main/resources/com/example/napolinights/SignUp.fxml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
-
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.image.Image?>
-<AnchorPane xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.napolinights.SignUpController"
-            style="-fx-background-color: #0e0e0e; -fx-padding: 20px; -fx-border-width: 1px; -fx-border-color: #111;">
-
-    <!-- Center Content -->
-    <VBox alignment="CENTER" spacing="20" AnchorPane.topAnchor="20" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0">
-
-        <!-- Logo -->
-        <ImageView fx:id="logoImageView" fitHeight="200" fitWidth="248.0" preserveRatio="true" style="-fx-border-color: red; -fx-border-width: 10px; -fx-padding: 10px;">
-            <image>
-                <Image url="@images/landing_page/napoli_nights_logo.png"/>
-            </image>
-        </ImageView>
-
-        <!-- Signup Form -->
-        <VBox spacing="15" alignment="CENTER" style="-fx-background-color: #1e1e1e; -fx-padding: 20px; -fx-border-color: lightgray;">
-
-            <Label fx:id="lblSignUpStatusMessage" visible="false" />
-
-            <!-- Input Fields -->
-            <TextField fx:id="emailField" promptText="Email address" style="-fx-background-color: #2e2e2e; -fx-text-fill: white; -fx-border-color: #555; -fx-padding: 10px;"/>
-            <Label fx:id="lblEmailMessage" text="* Required field is empty" visible="false"/>
-
-            <PasswordField fx:id="passwordField" promptText="Password (Minimum 6 characters)" style="-fx-background-color: #2e2e2e; -fx-text-fill: white; -fx-border-color: #555; -fx-padding: 10px;"/>
-            <Label fx:id="lblPasswordMessage" text="* Required field is empty" visible="false"/>
-
-            <TextField fx:id="firstNameField" promptText="First name" style="-fx-background-color: #2e2e2e; -fx-text-fill: white; -fx-border-color: #555; -fx-padding: 10px;"/>
-            <Label fx:id="lblFirstNameMessage" text="* Required field is empty" visible="false"/>
-
-            <TextField fx:id="lastNameField" promptText="Last name" style="-fx-background-color: #2e2e2e; -fx-text-fill: white; -fx-border-color: #555; -fx-padding: 10px;"/>
-            <Label fx:id="lblLastNameMessage" text="* Required field is empty" visible="false"/>
-
-            <TextField fx:id="mobileField" promptText="Mobile" style="-fx-background-color: #2e2e2e; -fx-text-fill: white; -fx-border-color: #555; -fx-padding: 10px;"/>
-
-            <!-- Signup Button -->
-            <Button text="Sign Up" style="-fx-background-color: red; -fx-text-fill: white; -fx-font-size: 16px; -fx-padding: 10px 20px;" onAction="#handleSignUp"/>
-
-            <!-- Login Link -->
-            <Hyperlink fx:id="lnkLogin" text="Already Signed Up? Login Here" style="-fx-text-fill: lightgray;" onAction="#handleLogin"/>
-
-        </VBox>
-    </VBox>
-
-</AnchorPane>
-
