--- src/main/resources/com/example/napolinights/StaffLandingPage.fxml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.image.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
-
-<AnchorPane xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.napolinights.StaffLandingPageController"
-            style="-fx-background-color: #141415; -fx-padding: 20px;">
-
-    <!-- VBox to center content -->
-    <VBox alignment="CENTER" spacing="20" AnchorPane.topAnchor="20" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0">
-
-        <!-- Napoli Nights Logo -->
-        <ImageView fx:id="logoImageView" fitHeight="100" fitWidth="150" preserveRatio="true">
-            <image>
-                <Image url="@images/landing_page/napoli_nights_logo.png"/>
-            </image>
-        </ImageView>
-
-        <!-- Admin Label -->
-        <Label text="ADMIN" style="-fx-font-size: 24px; -fx-text-fill: black;"/>
-
-        <!-- Buttons for View Orders and Reports -->
-        <HBox alignment="CENTER" spacing="40">
-            <Button fx:id="viewOrdersButton" text="View Incoming Orders" style="-fx-background-color: #ffffff; -fx-border-color: #000000; -fx-border-radius: 10px; -fx-font-size: 18px; -fx-text-fill: black;" prefWidth="200" prefHeight="100"/>
-            <Button fx:id="viewReportsButton" text="View Reports" style="-fx-background-color: #ffffff; -fx-border-color: #000000; -fx-border-radius: 10px; -fx-font-size: 18px; -fx-text-fill: black;" prefWidth="200" prefHeight="100"/>
-        </HBox>
-    </VBox>
-
-</AnchorPane>
