+++ src/main/resources/view/EmptyCartError.fxml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.text.*?>
+
+<VBox alignment="CENTER" prefHeight="93.0" prefWidth="242.0" spacing="10" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.napolinights.controller.EmptyCartErrorController">
+    <Text style="-fx-font-size: 18px;" text="Error: Your cart is empty!" />
+    <Button fx:id="closeButton" onAction="#handleClose" text="Close" />
+</VBox>
