+++ src/main/resources/view/ItemDetailDialog.fxml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.image.ImageView?>
+
+<VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" spacing="10" alignment="CENTER" fx:controller="com.example.napolinights.controller.ItemDetailDialogController">
+    <ImageView fx:id="dialogImageView" fitWidth="300" preserveRatio="true"/>
+    <Label fx:id="dialogTitle" style="-fx-font-size: 22px;"/>
+    <Label fx:id="dialogDescription" style="-fx-font-size: 14px; -fx-text-fill: gray;"/>
+    <Label fx:id="dialogPrice" style="-fx-font-size: 18px;"/>
+
+    <!-- Quantity Controls -->
+    <HBox spacing="10" alignment="CENTER">
+        <Button text="-" onAction="#decrementQuantity"/>
+        <Label fx:id="lblQuantity" text="1"/>
+        <Button text="+" onAction="#incrementQuantity"/>
+    </HBox>
+
+    <!-- Notes TextArea -->
+    <Label text="Notes"/>
+    <TextArea fx:id="notesTextArea" prefRowCount="3"/>
+
+    <!-- Footer Buttons -->
+    <HBox spacing="20" alignment="CENTER">
+        <Button fx:id="btnCancel" text="Cancel" onAction="#handleCancel"/>
+        <Button text="Add to Cart" onAction="#handleAddToCart"/>
+    </HBox>
+</VBox>
+
